Simple meta heuristics for projects management
In the real work there is a lot of constrain when we have to evaluate
if give a priority to one project or not.
We have to consider:
-  turnover
-  profitability
- timing and schedule
- required resources
- internal costs

The target is increase the revenue, the turnover, reduce the internal
costs and use as much as possible the resources.

The scope is to test different approach using Simulated Annealing,
Montecarlo, Taboo search and Genetic algorithm.
It will require too much time to me, sorry if is not completed